+++ server/httpServer.js
@@ -56,7 +56,7 @@ export const httpServerStart = function httpServerStart(port) {
 
 // uncomment after placing your favicon in /public
 //app.use(favicon(path.join(__dirname, 'public', 'favicon.ico')));
-  app.use(logger('dev'));
+  app.use(logger('tiny'));
   app.use(bodyParser.json());
   app.use(bodyParser.urlencoded({ extended: false }));
   app.use(cookieParser());